Use decimals instead of double

+++ DesktopClient.Services/AssetPriceMeasurement.cs
@@ -1,5 +1,5 @@
 using System;
 
 namespace DesktopClient.Services {
-	public record AssetPriceMeasurement(DateTime Date, double TotalPrice, double CumulativeFunds);
+	public record AssetPriceMeasurement(DateTime Date, decimal TotalPrice, decimal CumulativeFunds);
 }
