+++ DesktopClient.Services/ExchangeDto.cs
@@ -1,5 +1,5 @@
 using System;
 
 namespace InvestmentAnalyzer.State {
-	public record ExchangeDto(DateOnly Date, string CharCode, double Nominal, double Value);
+	public record ExchangeDto(DateOnly Date, string CharCode, decimal Nominal, decimal Value);
 }
