+++ State/Persistant/Exchange.cs
@@ -1,3 +1,3 @@
 namespace InvestmentAnalyzer.State {
-	public record Exchange(string Date, string CharCode, double Nominal, double Value);
+	public record Exchange(string Date, string CharCode, decimal Nominal, decimal Value);
 }
