+++ State/PortfolioStateEntry.cs
@@ -2,5 +2,5 @@ using System;
 
 namespace InvestmentAnalyzer.State {
 	public record PortfolioStateEntry(
-		DateOnly Date, string BrokerName, string ISIN, string Name, string Currency, double Count, double TotalPrice, double PricePerUnit);
+		DateOnly Date, string BrokerName, string ISIN, string Name, string Currency, decimal Count, decimal TotalPrice, decimal PricePerUnit);
 }
